requirements.txt
langchain==0.2.12
pandas==1.5.3
python-dotenv==1.0.1
